pyproject.toml
[project]
name = "aureus-backend"
version = "0.1.0"
description = "Backend for Aureus wealth tracker"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dlt[postgres] (>=1.10.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "cryptography (>=44.0.3,<45.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
exclude = ["dbt_packages/", "target/"]

[tool.ruff.lint]
extend-select = ["E501"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
